+++ server/src/metakb/repository/queries/__init__.py
@@ -1,0 +1,5 @@
+"""Provide access to static Neo4j queries."""
+
+from metakb.repository.queries.catalog import CypherCatalog
+
+__all__ = ["CypherCatalog"]
